[APPC-4627] Recover intent data on restart
There was an issue when the app was launched from a promocode or giftcard and there was no wallet. In this scenario the promocode/giftcard screens were not showing up with the data. Fix this issue by recovering the intent on restart. The onboarding is not recovering the intent because if the onboarding is appearing, its because the app was not started from a promocode/giftcard deeplink